requirements.txt
python-dotenv==0.20.0
interactions-fastapi==0.0.1